refactor Level2ProductLayer
